Move `find_nth_occurrence` to utils helpers
