+++ bot/utils/__init__.py
@@ -1,5 +1,5 @@
-from bot.utils.helpers import CogABCMeta, pad_base64
+from bot.utils.helpers import CogABCMeta, find_nth_occurrence, pad_base64
 from bot.utils.redis_cache import RedisCache
 from bot.utils.services import send_to_paste_service
 
-__all__ = ['RedisCache', 'CogABCMeta', "pad_base64", "send_to_paste_service"]
+__all__ = ['RedisCache', 'CogABCMeta', 'find_nth_occurrence', 'pad_base64', 'send_to_paste_service']
