+++ bot/utils/helpers.py
@@ -1,10 +1,21 @@
 from abc import ABCMeta
+from typing import Optional
 
 from discord.ext.commands import CogMeta
 
 
 class CogABCMeta(CogMeta, ABCMeta):
     """Metaclass for ABCs meant to be implemented as Cogs."""
+
+
+def find_nth_occurrence(string: str, substring: str, n: int) -> Optional[int]:
+    """Return index of `n`th occurrence of `substring` in `string`, or None if not found."""
+    index = 0
+    for _ in range(n):
+        index = string.find(substring, index+1)
+        if index == -1:
+            return None
+    return index
 
 
 def pad_base64(data: str) -> str:
